requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
torch==2.2.0
torchaudio==2.2.0
transformers==4.37.2
pyannote.audio==3.1.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0 
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
python-multipart==0.0.6